# docs/swagger/index.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "room_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "RoomService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/rooms": {
      "post": {
        "operationId": "RoomService_CreateRoom",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateRoomResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateRoomRequest"
            }
          }
        ],
        "tags": [
          "RoomService"
        ]
      }
    },
    "/v1/rooms/join/{id}": {
      "post": {
        "operationId": "RoomService_JoinRoom",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbJoinRoomResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "RoomService"
        ]
      }
    }
  },
  "definitions": {
    "pbCreateRoomRequest": {
      "type": "object",
      "properties": {
        "owner": {
          "type": "string"
        },
        "roomSize": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "pbCreateRoomResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "roomName": {
          "type": "string"
        },
        "roomSize": {
          "type": "integer",
          "format": "int64"
        },
        "createdBy": {
          "type": "string"
        }
      }
    },
    "pbJoinRoomResponse": {
      "type": "object",
      "properties": {
        "roomId": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}